+++ packages/core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @bundleup/core
 
+## 0.0.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @bundleup/sdk@0.0.14
+
 ## 0.0.11
 
 ### Patch Changes