+++ packages/nextjs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @bundleup/nextjs
 
+## 0.0.14
+
+### Patch Changes
+
+- @bundleup/core@0.0.12
+- @bundleup/react@0.0.14
+
 ## 0.0.13
 
 ### Patch Changes