+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @bundleup/sdk
 
+## 0.0.14
+
+### Patch Changes
+
+- added unified methods
+
 ## 0.0.13
 
 ### Patch Changes